Stop using constrainMaxWidth prop.
